Update TheNoobBot to 5.0.4. Sorry for CRLF/LF switching :o
Note: Others UnitName (players) is not retrieved,
Quester offset to update,
WoWGameObject.GOType to update
Pandaren for Alliance&Horde to add in the Enum (neutral pandaren done)
Spelllist.txt needs update
Listmount.txt needs update
Nodes.xml needs update (if needed)
Need to check if Skillline (? skill level) is working properly.

+++ The Noob Bot/nManager/Wow/MemoryClass/D3D.cs
@@ -60,10 +60,10 @@ namespace nManager.Wow.MemoryClass
         public static uint D3D9Adresse(int processId)
         {
             var memory = new Magic.BlackMagic(processId);
-            uint pDevice = memory.ReadUInt((uint)memory.GetModule("Wow.exe").BaseAddress + Patchables.Addresses.Hooking.DX_DEVICE);
-            uint pEnd = memory.ReadUInt(pDevice + Patchables.Addresses.Hooking.DX_DEVICE_IDX);
+            uint pDevice = memory.ReadUInt((uint)memory.GetModule("Wow.exe").BaseAddress + (uint)Patchables.Addresses.Hooking.DX_DEVICE);
+            uint pEnd = memory.ReadUInt(pDevice + (uint)Patchables.Addresses.Hooking.DX_DEVICE_IDX);
             uint pScene = memory.ReadUInt(pEnd);
-            d3d9Adresse = memory.ReadUInt(pScene + Patchables.Addresses.Hooking.ENDSCENE_IDX);
+            d3d9Adresse = memory.ReadUInt(pScene + (uint)Patchables.Addresses.Hooking.ENDSCENE_IDX);
 
             return d3d9Adresse;
         }
